Implement get in DynamoDB service
Implement getByUserIdAndCode;

TODO:
- break backend up into a multi-module project:
  - a project per function;
  - a project for the DAOs/Services;
  - a project for commons;
- add security / Cognito integration to put requests;
- implement extra filtering options in the DynamoDB services;
- add logging;
- decouple request validation.

+++ backend/ConsultFunction/src/main/java/com/lucaspetrini/consult/request/PutUserRatingRequest.java
@@ -2,16 +2,16 @@ package com.lucaspetrini.consult.request;
 
 public class PutUserRatingRequest {
 
-	private Integer rating;
+	private Long rating;
 	private Long date;
 	private String review;
 
 
-	public void setRating(Integer rating) {
+	public void setRating(Long rating) {
 		this.rating = rating;
 	}
 	
-	public Integer getRating() {
+	public Long getRating() {
 		return rating;
 	}
