+++ backend/ConsultFunction/src/test/java/com/lucaspetrini/consult/mapper/JsonObjectMapperTest.java
@@ -14,7 +14,7 @@ import com.lucaspetrini.consult.response.GetUserRatingResponse;
 public class JsonObjectMapperTest {
 	private static final String SKU = "SKU123";
 	private static final String USER = "mrpickles";
-	private static final Integer RATING = 3;
+	private static final Long RATING = 3L;
 	private static final String REVIEW = "Good";
 	private static final String SERIALISED_PUT_REQUEST = "{'rating':3,'date':9941,'review':'Good'}".replace("'", "\"");
 	private static final String DESERIALISED_GET_RESPONSE = "{'sku':'SKU123','user':'mrpickles','rating':3,'date':9941,'review':'Good'}".replace("'", "\"");
